Non-self-hosted code was out of sync.
When `EE_CFG_SELFHOSTED != 1`, the harness initialized the
the timestamp the wrong way and did not supress all verification
timestamps, which cause failures in both GUI runner programs.

The code has been reverted back to the expected behavior and
validated on the both current versions of (license) GUIs.

* Version 1.0.2.000 energy runner

* Version 1.0.2.001 performance runner

Both GUIs should generate 45 timestamps in both configurations,
`EE_CFG_ENERGY_MODE` set to `0` and `1` with the GUI scripts.