Fixed confusion with g_verify_mode and timestamps.
Created a new variable called g_mute_timestamps which
is a consequence of setting verify mode via the parser.
It got a little confusing since `verify` means to print
out verifcation messages that the host checks. Compared
to muting timestamps, which is required when verifying,
but also when doing a bench encrypt before decrypt.